+++ services/api-server/src/simcore_service_api_server/utils/client_decorators.py
@@ -24,7 +24,7 @@ from tenacity import (
     wait_fixed,
 )
 
-from ..models.raw_data import JSON
+from ..models.types import JSON
 
 
 def handle_retry(logger: logging.Logger):
